requirements.txt
mujoco==2.2.0
gym==0.25.2
tensorboard==2.10.0
tensorboardX==2.5.1
matplotlib==3.5.3
ipython==7.34.0
moviepy==1.0.3
pyvirtualdisplay==3.0
torch==1.12.1
opencv-python==4.6.0.66
ipdb==0.13.9
swig==4.0.2
box2d-py==2.3.8
jax==0.2.20
jaxlib
wandb==0.12.5
torchvision>=0.6
numpy
termcolor
packaging
gym[atari]
gpustat
pygame
dm_haiku
